Separate ApiUpgradeHandler from the manager
