+++ upgrader/src/main/java/org/gradle/demo/api/evolution/ApiUpgradeHandler.java
@@ -1,0 +1,35 @@
+package org.gradle.demo.api.evolution;
+
+import com.google.common.collect.ImmutableList;
+import org.codehaus.groovy.runtime.callsite.CallSite;
+import org.codehaus.groovy.runtime.callsite.CallSiteArray;
+
+public class ApiUpgradeHandler {
+    private static ApiUpgradeHandler INSTANCE;
+
+    private final ImmutableList<Replacement> replacements;
+
+    public ApiUpgradeHandler(ImmutableList<Replacement> replacements) {
+        this.replacements = replacements;
+    }
+
+    public void useInstance() {
+        INSTANCE = this;
+    }
+
+    @SuppressWarnings("unchecked")
+    public static <T> T invokeReplacement(Object receiver, Object[] args, int methodReplacementIndex) {
+        MethodReplacement<T> methodReplacement = (MethodReplacement<T>) INSTANCE.replacements.get(methodReplacementIndex);
+        return (T) methodReplacement.invokeReplacement(receiver, args);
+    }
+
+    public static void decorateCallSiteArray(CallSiteArray callSites) {
+        for (CallSite callSite : callSites.array) {
+            for (Replacement replacement : INSTANCE.replacements) {
+                replacement.decorateCallSite(callSite).ifPresent(decoreated ->
+                    callSites.array[callSite.getIndex()] = decoreated
+                );
+            }
+        }
+    }
+}
